#41 refactor: move dto in each packages
