--- Ppap/app/src/main/java/com/jeongg/ppap/data/dto/NoticeDTO.kt
@@ -1,13 +1,0 @@
-package com.jeongg.ppap.data.dto
-
-import kotlinx.datetime.LocalDateTime
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class NoticeDTO (
-    val contentId: Long,
-    val title: String,
-    val pubDate: LocalDateTime,
-    val link: String,
-    var isScraped: Boolean
-)
