--- Ppap/app/src/main/java/com/jeongg/ppap/data/dto/SubscribeDTO.kt
@@ -1,9 +1,0 @@
-package com.jeongg.ppap.data.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class SubscribeDTO(
-    val subscribeId: Long,
-    val title: String
-)
