+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/notice/NoticeDTO.kt
@@ -1,13 +1,13 @@
-package com.jeongg.ppap.data.dto
+package com.jeongg.ppap.data.dto.notice
 
 import kotlinx.datetime.LocalDateTime
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ScrapDTO(
+data class NoticeDTO (
     val contentId: Long,
-    val contentTitle: String,
-    val link: String,
+    val title: String,
     val pubDate: LocalDateTime,
-    var isScrap: Boolean
+    val link: String,
+    var isScraped: Boolean
 )
