+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/notice/NoticeItemDTO.kt
@@ -1,4 +1,4 @@
-package com.jeongg.ppap.data.dto
+package com.jeongg.ppap.data.dto.notice
 
 import androidx.compose.runtime.MutableState
