+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/scrap/ScrapDTO.kt
@@ -1,0 +1,13 @@
+package com.jeongg.ppap.data.dto.scrap
+
+import kotlinx.datetime.LocalDateTime
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ScrapDTO(
+    val contentId: Long,
+    val contentTitle: String,
+    val link: String,
+    val pubDate: LocalDateTime,
+    var isScrap: Boolean
+)
