+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/subscribe/SubscribeGetResponseDTO.kt
@@ -1,10 +1,10 @@
-package com.jeongg.ppap.data.dto
+package com.jeongg.ppap.data.dto.subscribe
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class UnivNoticeBoardDTO(
+data class SubscribeGetResponseDTO(
     val subscribeId: Long,
     val title: String,
-    val link: String
+    var isActive: Boolean
 )
