+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/subscribe/SubscribeUpdateRequestDTO.kt
@@ -1,4 +1,4 @@
-package com.jeongg.ppap.data.dto
+package com.jeongg.ppap.data.dto.subscribe
 
 import kotlinx.serialization.Serializable
