+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/subscribe/SubscribeUpdateResponseDTO.kt
@@ -1,12 +1,11 @@
-package com.jeongg.ppap.data.dto
+package com.jeongg.ppap.data.dto.subscribe
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SubscribeWithContentDTO(
+data class SubscribeUpdateResponseDTO(
     val subscribeId: Long = 0,
     val title: String = "",
     val noticeLink: String? = null,
-    val rssLink: String = "",
     val isActive: Boolean = false
 )
