+++ Ppap/app/src/main/java/com/jeongg/ppap/data/dto/subscribe/SubscribeWithContentDTO.kt
@@ -1,11 +1,12 @@
-package com.jeongg.ppap.data.dto
+package com.jeongg.ppap.data.dto.subscribe
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SubscribeUpdateResponseDTO(
+data class SubscribeWithContentDTO(
     val subscribeId: Long = 0,
     val title: String = "",
     val noticeLink: String? = null,
+    val rssLink: String = "",
     val isActive: Boolean = false
 )
