+++ Ppap/app/src/main/java/com/jeongg/ppap/data/util/PError.kt
@@ -1,6 +1,5 @@
 package com.jeongg.ppap.data.util
 
-import io.ktor.client.call.NoTransformationFoundException
 import io.ktor.client.plugins.ClientRequestException
 import io.ktor.client.plugins.RedirectResponseException
 import io.ktor.client.plugins.ServerResponseException
