+++ Ppap/app/src/main/java/com/jeongg/ppap/domain/usecase/subscribe/GetUnivSubscribeList.kt
@@ -1,6 +1,6 @@
 package com.jeongg.ppap.domain.usecase.subscribe
 
-import com.jeongg.ppap.data.dto.UnivNoticeBoardDTO
+import com.jeongg.ppap.data.dto.univ.UnivNoticeBoardDTO
 import com.jeongg.ppap.data.util.ApiUtils
 import com.jeongg.ppap.data.util.getErrorMessage
 import com.jeongg.ppap.domain.repository.SubscribeRepository
