+++ Ppap/app/src/main/java/com/jeongg/ppap/domain/usecase/subscribe/UpdateSubscribe.kt
@@ -1,7 +1,7 @@
 package com.jeongg.ppap.domain.usecase.subscribe
 
-import com.jeongg.ppap.data.dto.SubscribeUpdateRequestDTO
-import com.jeongg.ppap.data.dto.SubscribeUpdateResponseDTO
+import com.jeongg.ppap.data.dto.subscribe.SubscribeUpdateRequestDTO
+import com.jeongg.ppap.data.dto.subscribe.SubscribeUpdateResponseDTO
 import com.jeongg.ppap.data.util.ApiUtils
 import com.jeongg.ppap.data.util.getErrorMessage
 import com.jeongg.ppap.domain.repository.SubscribeRepository
