+++ Ppap/app/src/main/java/com/jeongg/ppap/presentation/state/NoticeItemState.kt
@@ -1,8 +1,8 @@
 package com.jeongg.ppap.presentation.state
 
 import androidx.paging.PagingData
-import com.jeongg.ppap.data.dto.NoticeItemDTO
-import com.jeongg.ppap.data.dto.SubscribeGetResponseDTO
+import com.jeongg.ppap.data.dto.notice.NoticeItemDTO
+import com.jeongg.ppap.data.dto.subscribe.SubscribeGetResponseDTO
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flow
