--- Ppap/app/src/main/java/com/jeongg/ppap/util/PLog.kt
@@ -1,7 +1,0 @@
-package com.jeongg.ppap.util
-
-import android.util.Log
-
-fun String.log(){
-    Log.d("ppap_log", this)
-}
